Refactored the location of some packages
