+++ src/GUIClasses/ActionListeners/ProctorView/DormitoryView/BackLabelListener.java
@@ -1,4 +1,4 @@
-package GUIClasses.ActionListeners.DormitoryView;
+package GUIClasses.ActionListeners.ProctorView.DormitoryView;
 
 import GUIClasses.ProctorViews.DormitoryView;
