+++ src/GUIClasses/ActionListeners/ProctorView/ProctorPage/ReportMenuItemListener.java
@@ -1,4 +1,4 @@
-package GUIClasses.ActionListeners.ProctorPage;
+package GUIClasses.ActionListeners.ProctorView.ProctorPage;
 
 import GUIClasses.ProctorViews.ProctorPage;
 import GUIClasses.ProctorViews.ReportsView;
